feat: update logic for music and update word_set

+++ application/main.py
@@ -1,6 +1,6 @@
 from crawler.data_crawl_process import get_mp3_url_song
 from utils.info_extraction import info_extractor, normal_str
-from utils.iot_process import send_signal
+from utils.iot_process import send_signal, send_url
 from utils.prediction import get_result_request
 import firebase_admin
 from firebase_admin import db, credentials
@@ -17,16 +17,19 @@ def listener(event):
     print(event.data)  # new data at /reference/event.path. None if deleted
     # get_result_request(event.data)
     # device, status, room, color_code, song_name = get_intent_info("Đóng cửa phòng khách cho tôi")
-    intent, status, room, color_code, song_name = get_intent_info(event.data)
-    print("intent: ", intent)
-    if intent == 'music':
-        print("song_name: ", normal_str(song_name))
-        print("url: ", get_mp3_url_song(normal_str(song_name)))
-    else:
-        send_signal(room, intent, color_code, status)
-        print("status: ", status)
-        print("room: ", room)
-        print("color_code: ", color_code)
+    if event.data != "":
+        intent, status, room, color_code, song_name = get_intent_info(event.data)
+        print("intent: ", intent)
+
+        if intent == 'music':
+            send_url(get_mp3_url_song(normal_str(song_name)))
+            print("song_name: ", normal_str(song_name))
+            print("url: ", get_mp3_url_song(normal_str(song_name)))
+        else:
+            send_signal(room, intent, color_code, status)
+            print("status: ", status)
+            print("room: ", room)
+            print("color_code: ", color_code)
 
 
 json_path = r'C:\Users\Public\cred.json'
